requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.7.2
